Restored warping and mhx progress bar after alpha 7.
git-svn-id: http://makehuman.googlecode.com/svn@3813 ea038faa-237f-11de-8bc4-bb72633a7edc

+++ trunk/makehuman/core/warp.py
@@ -276,8 +276,8 @@ def compute_distance2(x, y=None):
 class CWarp2(object):
     
     def __init__(self, source, target, landmarks):
-        self.source = numpy.asarray(source, dtype="float")
-        self.target = numpy.asarray(target, dtype="float")
+        self.source = numpy.asarray(source, dtype="float32")
+        self.target = numpy.asarray(target, dtype="float32")
         
         self.xverts = self.source[landmarks]
         self.yverts = self.target[landmarks]
